Added oi parameter in gainers losers api
Parameter named "ret_oi" added which when set as true,
returns the gainers and losers based on Open Interest
difference in contrast to the close price difference.

"ret_oi" set to true works only when "type" is set to
"futures" as in the cases of "stocks" or "indices", the
oi is irrelevant.

Eg:
1.
{
    "number":5,
    "gainers_or_losers":"gainers",
    "type":"futures",
    "expiry_date":"2021-07-29",
    "ret_oi":true
}
Returns top 5 gainers in futures oi

2.
{
    "number":5,
    "gainers_or_losers":"losers",
    "type":"futures",
    "expiry_date":"2021-07-29",
}
Returns top 5 losers in futures (in terms of close price)

3.
{
    "number":5,
    "gainers_or_losers":"both",
    "type":"indices",
    "expiry_date":"2021-07-29",
    "ret_oi":true
}
Returns top 5 gainers & losers of indices
(in terms of close price because oi is irrelevant)