Fixed Cash Futures Arbitrage Api
The ltp of the stock prices were off by one entry due to that entry
not being recognized by the quote api.

Hence fixed the concatenation of the quote last prices by doing inner
join based on the column instead of the pasting the entries
as was the case earlier.